try split up things
