+++ packages/next/src/server/lib/server-ipc/index.ts
@@ -1,9 +1,9 @@
-import type NextServer from '../next-server'
-import { genExecArgv, getNodeOptionsWithoutInspect } from './utils'
-import { deserializeErr, errorToJSON } from '../render'
-import { IncomingMessage } from 'http'
+import type NextServer from '../../next-server'
+
+import { genExecArgv, getNodeOptionsWithoutInspect } from '../utils'
+import { deserializeErr, errorToJSON } from '../../render'
 import crypto from 'crypto'
-import isError from '../../lib/is-error'
+import isError from '../../../lib/is-error'
 
 // we can't use process.send as jest-worker relies on
 // it already and can cause unexpected message errors
@@ -88,7 +88,7 @@ export const createWorker = (
 ) => {
   const { initialEnv } = require('@next/env') as typeof import('@next/env')
   const { Worker } = require('next/dist/compiled/jest-worker')
-  const worker = new Worker(require.resolve('./render-server'), {
+  const worker = new Worker(require.resolve('../render-server'), {
     numWorkers: 1,
     // TODO: do we want to allow more than 10 OOM restarts?
     maxRetries: 10,
@@ -125,94 +125,13 @@ export const createWorker = (
       'clearModuleContext',
     ],
   }) as any as InstanceType<typeof Worker> & {
-    initialize: typeof import('./render-server').initialize
-    deleteCache: typeof import('./render-server').deleteCache
-    deleteAppClientCache: typeof import('./render-server').deleteAppClientCache
+    initialize: typeof import('../render-server').initialize
+    deleteCache: typeof import('../render-server').deleteCache
+    deleteAppClientCache: typeof import('../render-server').deleteAppClientCache
   }
 
   worker.getStderr().pipe(process.stderr)
   worker.getStdout().pipe(process.stdout)
 
   return worker
-}
-
-const forbiddenHeaders = [
-  'accept-encoding',
-  'content-length',
-  'keepalive',
-  'content-encoding',
-  'transfer-encoding',
-  // https://github.com/nodejs/undici/issues/1470
-  'connection',
-]
-
-export const filterReqHeaders = (
-  headers: Record<string, undefined | string | string[]>
-) => {
-  for (const [key, value] of Object.entries(headers)) {
-    if (
-      forbiddenHeaders.includes(key) ||
-      !(Array.isArray(value) || typeof value === 'string')
-    ) {
-      delete headers[key]
-    }
-  }
-  return headers
-}
-
-export const invokeRequest = async (
-  targetUrl: string,
-  requestInit: {
-    headers: IncomingMessage['headers']
-    method: IncomingMessage['method']
-  },
-  readableBody?: import('stream').Readable
-) => {
-  const parsedUrl = new URL(targetUrl)
-
-  // force localhost to IPv4 as some DNS may
-  // resolve to IPv6 instead
-  if (parsedUrl.hostname === 'localhost') {
-    parsedUrl.hostname = '127.0.0.1'
-  }
-  const invokeHeaders = filterReqHeaders({
-    ...requestInit.headers,
-  }) as IncomingMessage['headers']
-
-  const invokeRes = await new Promise<IncomingMessage>(
-    (resolveInvoke, rejectInvoke) => {
-      const http = require('http') as typeof import('http')
-
-      try {
-        const invokeReq = http.request(
-          targetUrl,
-          {
-            headers: invokeHeaders,
-            method: requestInit.method,
-          },
-          (res) => {
-            resolveInvoke(res)
-          }
-        )
-        invokeReq.on('error', (err) => {
-          rejectInvoke(err)
-        })
-
-        if (requestInit.method !== 'GET' && requestInit.method !== 'HEAD') {
-          if (readableBody) {
-            readableBody.pipe(invokeReq)
-            readableBody.on('close', () => {
-              invokeReq.end()
-            })
-          }
-        } else {
-          invokeReq.end()
-        }
-      } catch (err) {
-        rejectInvoke(err)
-      }
-    }
-  )
-
-  return invokeRes
 }
