+++ packages/next/src/server/next-server.ts
@@ -106,7 +106,8 @@ import { getRouteRegex } from '../shared/lib/router/utils/route-regex'
 import { removePathPrefix } from '../shared/lib/router/utils/remove-path-prefix'
 import { addPathPrefix } from '../shared/lib/router/utils/add-path-prefix'
 import { pathHasPrefix } from '../shared/lib/router/utils/path-has-prefix'
-import { filterReqHeaders, invokeRequest } from './lib/server-ipc'
+import { invokeRequest } from './lib/server-ipc/invoke-request'
+import { filterReqHeaders } from './lib/server-ipc/utils'
 import { createRequestResponseMocks } from './lib/mock-request'
 
 export * from './base-server'
